+++ wordpress-plugin-boilerplate-light/wordpress-plugin-boilerplate-light.php
@@ -3,7 +3,7 @@
  * Plugin Name:       WordPress Plugin Boilerplate Light
  * Plugin URI:        https://github.com/BoilWP/WordPress-Plugin-Boilerplate-Light
  * Description:       The best light WordPress plugin boilerplate you will ever need, comes prepaired with settings control panel.
- * Version:           1.0.0
+ * Version:           1.0.1
  * Author:            Sébastien Dumont
  * Author URI:        http://www.sebastiendumont.com
  * License:           GPL-2.0+
@@ -90,7 +90,7 @@ final class Plugin_Name {
 	 * @access public
 	 * @var    string
 	 */
-	public $version = "1.0.0";
+	public $version = "1.0.1";
 
 	/**
 	 * The WordPress version the plugin requires minumum.
@@ -125,34 +125,31 @@ final class Plugin_Name {
 	/**
 	 * The WordPress.org Plugin URI.
 	 *
-	 * @todo    Replace 'your-plugin-name' with the name of the
-	 *          plugin slug given for your wordpress repository
+	 * @todo    Replace 'plugin-name' with the name of the plugin slug given for your wordpress repository.
 	 * @since   1.0.0
-	 * @example https://wordpress.org/plugins/your-plugin-name
+	 * @example https://wordpress.org/plugins/plugin-name
 	 * @access  public
 	 * @var     string
 	 */
-	public $wp_plugin_url = "https://wordpress.org/plugins/your-plugin-name";
+	public $wp_plugin_url = "https://wordpress.org/plugins/plugin-name";
 
 	/**
 	 * The WordPress.org Plugin Support URI.
 	 *
-	 * @todo    Replace 'your-plugin-name' with the name of the
-	 *          plugin slug given for your wordpress repository
+	 * @todo    Replace 'plugin-name' with the name of the plugin slug given for your wordpress repository.
 	 * @since   1.0.0
 	 * @example https://wordpress.org/support/plugin/your-plugin-name
 	 * @access  public
 	 * @var     string
 	 */
-	public $wp_plugin_support_url = "https://wordpress.org/support/plugin/your-plugin-name";
+	public $wp_plugin_support_url = "https://wordpress.org/support/plugin/plugin-name";
 
 	/**
 	 * The WordPress.org Plugin Review URI.
 	 *
-	 * @todo    Replace 'your-plugin-name' with the name of the
-	 *          plugin slug given for your wordpress repository
-	 * @since   1.0.2
-	 * @example https://wordpress.org/support/view/plugin-reviews/your-plugin-name
+	 * @todo    Replace 'plugin-name' with the name of the plugin slug given for your wordpress repository.
+	 * @since   1.0.0
+	 * @example https://wordpress.org/support/view/plugin-reviews/plugin-name
 	 * @access  public
 	 * @var     string
 	 */
@@ -378,7 +375,7 @@ final class Plugin_Name {
 	 * Define Constants
 	 *
 	 * @todo   1. Change 'PLUGIN_NAME' to the name of the plugin.
-	 *         2. Change 'plugin-name' with the plugin slug of your plugin on "WordPress.org"
+	 * @todo   2. Change 'plugin-name' with the plugin slug of your plugin on "WordPress.org"
 	 * @since  1.0.0
 	 * @access private
 	 */
@@ -399,7 +396,7 @@ final class Plugin_Name {
 		if ( ! defined( 'PLUGIN_NAME_GITHUB_REPO_URI' ) )       define( 'PLUGIN_NAME_GITHUB_REPO_URI', $this->github_repo_url );
 
 		$suffix = defined( 'SCRIPT_DEBUG' ) && SCRIPT_DEBUG ? '' : '.min';
-		define( 'PLUGIN_NAME_SCRIPT_MODE', $suffix );
+		if ( ! defined( 'PLUGIN_NAME_SCRIPT_MODE' ) )           define( 'PLUGIN_NAME_SCRIPT_MODE', $suffix );
 	} // END define_constants()
 
 	/**
@@ -490,8 +487,7 @@ final class Plugin_Name {
 	 */
 	public function frontend_includes() {
 		// Functions
-		include_once( 'includes/plugin-name-template-hooks.php' ); // Include template hooks for themes to remove/modify them
-		include_once( 'includes/plugin-name-functions.php' );      // Contains functions for various front-end events
+		include_once( 'includes/plugin-name-functions.php' ); // Contains functions for various front-end events
 	} // END frontend_includes()
 
 	/**
@@ -499,7 +495,6 @@ final class Plugin_Name {
 	 *
 	 * @since  1.0.0
 	 * @access public
-	 * @filter plugin_name_force_ssl_filter
 	 */
 	public function init_plugin_name() {
 		// Before init action
@@ -579,18 +574,6 @@ final class Plugin_Name {
 	} // END plugin_path()
 
 	/**
-	 * Get the plugin template path.
-	 *
-	 * @since  1.0.0
-	 * @access public
-	 * @filter plugin_name_template_path
-	 * @return string
-	 */
-	public function template_path() {
-		return apply_filters( 'plugin_name_template_path', 'plugin-name/' );
-	} // END template_path()
-
-	/**
 	 * Registers and enqueues stylesheets and javascripts
 	 * for the administration panel and the front of the site.
 	 *
@@ -656,8 +639,8 @@ final class Plugin_Name {
 	 *
 	 * @since  1.0.0
 	 * @access private
-	 * @param  string  $name	    The ID to register with WordPress.
-	 * @param  string  $file_path	The path to the actual file.
+	 * @param  string  $name      The ID to register with WordPress.
+	 * @param  string  $file_path The path to the actual file.
 	 * @param  bool    $is_script Optional, argument for if the incoming file_path is a JavaScript source file.
 	 * @param  array   $support   Optional, for requiring other javascripts for the source file you are calling.
 	 * @param  string  $version   Optional, can match the version of the plugin or version of the source file.
@@ -681,6 +664,7 @@ final class Plugin_Name {
 		} // end if
 
 		wp_enqueue_style( 'wp-color-picker' );
+
 		if ( is_admin() && $wp_version >= '3.8' ) {
 			wp_enqueue_style( 'dashicons' ); // Loads only in WordPress 3.8 and up.
 		}