+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.0.1 (22nd March 2015)
+
+* Cleaned the boilerplate more from forking "[WordPress Plugin Boilerplate](https://github.com/BoilWP/WordPress-Plugin-Boilerplate)"
+* Added the uninstall option in the settings.
+* Improved inline documentation.
+
 ## 1.0.0 (12th March 2015)
 
 * Initial Release